refactor: segregate controllers into V1, V2, and V3 components for optimization strategies
1. Architectural Refactoring (Version Segregation)
- Extract GameCharacterControllerV1: Isolate stability-focused logic (Pessimistic Lock & Consistency)
- Extract GameCharacterControllerV2: Isolate performance-focused logic (Caffeine Cache & Write-Behind pattern)
- Extract GameCharacterControllerV3: Isolate resource-optimization logic (Streaming & GZIP compression)

2. Documentation & Maintainability
- Add Architectural Javadocs: Document design intent, strategies, and trade-offs (e.g., OOM prevention, TTFB reduction) for each version
- Apply SRP: Decouple specific optimization strategies from the monolithic controller to improve readability
